requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
dj-database-url==2.2.0
Django==5.1.1
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.0
future==1.0.0
gevent==24.2.1
greenlet==3.1.0
gunicorn==23.0.0
idna==3.9
packaging==24.1
psycopg==3.2.1
psycopg2==2.9.9
PyJWT==2.9.0
PySocks==1.7.1
requests==2.32.3
requests-file==2.1.0
sqlparse==0.5.1
tldextract==5.1.2
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
zope.event==5.0
zope.interface==7.0.3
